fix(ml-thirdparty-simulator): refactor factories
