+++ src/shared/ml-thirdparty-simulator/factories/participant.ts
@@ -28,23 +28,46 @@ import * as faker from 'faker'
 import { Participant } from '~/shared/ml-thirdparty-client/models/core'
 
 /**
- * Number of participants that will be generated by the simulator.
- * 
- * TODO: customize the number of participants using the configs passed
- * to the Mojaloop simulator upon initialization.
+ * A class that helps to generate random participants for the simulator.
  */
-const numberOfParticipants = 10
+export class ParticipantFactory {
+  /**
+   * Number of participants that will be generated by the simulator. The value
+   * must be a positive integer. If not set, the default value is 10.
+   * 
+   * Note that this variable must be set before `getParticipants()` is called, 
+   * otherwise, changing the value will not have any effect since the list of 
+   * participants is stored within the `participants` field.
+   */
+  public static numOfParticipants: number = 10
 
-let participants: Participant[] = []
+  /**
+   * List of participants that is stored internally within the `ParticipantFactory`.
+   */
+  private static participants: Participant[] = []
 
-for (var i = 0; i < numberOfParticipants; i++) {
-  // Generate a random participant
-  let participant: Participant = {
-    fspId: faker.finance.bic(),
-    name: faker.company.companyName()
+  /**
+   * Returns the list of participants for the simulator.
+   * The participants will only be generated once when this function is called
+   * for the first time. Afterward, the list will be stored internally for 
+   * subsequent usage.
+   */
+  public static getParticipants() {
+    if (ParticipantFactory.participants.length < 1) {
+      ParticipantFactory.createParticipants()
+    }
+    return ParticipantFactory.participants
   }
 
-  participants.push(participant)
-}
+  private static createParticipants() {
+    for (var i = 0; i < ParticipantFactory.numOfParticipants; i++) {
+      // Generate a random participant
+      let participant: Participant = {
+        fspId: faker.finance.bic(),
+        name: faker.company.companyName()
+      }
 
-export { participants }
+      ParticipantFactory.participants.push(participant)
+    }
+  }
+}
