+++ test/unit/shared/ml-thirdparty-simulator/index.test.ts
@@ -29,7 +29,7 @@ import config from '~/lib/config'
 
 import { Simulator } from '~/shared/ml-thirdparty-simulator'
 import { PartyIdType } from '~/shared/ml-thirdparty-client/models/core'
-import { createPutPartiesRequest } from '~/shared/ml-thirdparty-simulator/factories'
+import { PartyFactory } from '~/shared/ml-thirdparty-simulator/factories/party'
 
 jest.useFakeTimers()
 
@@ -42,7 +42,7 @@ const partyLookupParams = {
 }
 
 jest.mock('~/lib/firebase')
-jest.mock('~/shared/ml-thirdparty-simulator/factories/parties')
+jest.mock('~/shared/ml-thirdparty-simulator/factories/party')
 
 describe('Mojaloop third-party simulator', () => {
   let simulator: Simulator
@@ -51,7 +51,7 @@ describe('Mojaloop third-party simulator', () => {
   beforeAll(async () => {
     server = await createServer(config)
     simulator = new Simulator(server, {
-      host: 'mojaloop.pisp-demo-server.local',
+      host: 'mojaloop.' + config.get('hostname'),
       delay: 100,
     })
   })
@@ -70,14 +70,14 @@ describe('Mojaloop third-party simulator', () => {
     await simulator.getParties(partyLookupParams.type, partyLookupParams.id)
 
     // payload that is injected to the server must match the one generated by the simulator
-    const payload = createPutPartiesRequest(partyLookupParams.type, partyLookupParams.id)
+    const payload = PartyFactory.createPutPartiesRequest(partyLookupParams.type, partyLookupParams.id)
 
     expect(serverSpy).toBeCalledTimes(1)
     expect(serverSpy).toBeCalledWith({
       method: 'PUT',
       url: targetUrl,
       headers: {
-        host: 'mojaloop.pisp-demo-server.local',
+        host: 'mojaloop.' + config.get('hostname'),
         'Content-Length': JSON.stringify(payload).length.toString(),
         'Content-Type': 'application/json',
       },
